fix theme editor save bug
This commit fixes a bug where the theme editor was not saving
an edited theme when the user clicked the save button. The
primary issue was session clearing logic on the login page
(which was being loaded into the theme editor via an iframe).

closes FOO-4105
flag=none

test plan:
LOGIN PAGE:
1. log-in to Canvas
2. open web dev toolbar
3. open Application tab and click on Session Storage
     - you should see an entry for a `QUERY_CACHE` key
4. log-out of Canvas
5. EXPECT Session Storage to be cleared

THEME EDITOR:
1. log-in to Canvas
2. open web dev toolbar
3. open Application tab and click on Session Storage
4. Navigate to an account’s theme page
5. create a new theme based off of the default template
6. make a change to the theme (color change will suffice)
7. click the save button, give the new theme a name, save
9. exit the theme editor
10. re-open the new theme in the theme editor
11. EXPECT `sharedBrandConfigBeingEdited` key to be present
    in the Session Storage
      - prior to this commit, the key was being nuked when
        the login page was iframe-loaded into the theme editor
12. make a change and preview
14. EXPECT `sharedBrandConfigBeingEdited` key to still be present
15. click the save button
13. EXPECT the “theme name” modal to NOT appear
15. exit theme editor
16. re-open the new theme in the theme editor
15. EXPECT theme to load changes from last edit/save

Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/334427
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
QA-Review: Michael Hulse <michael.hulse@instructure.com>
Product-Review: Michael Hulse <michael.hulse@instructure.com>

+++ ui/features/login/index.ts
@@ -114,5 +114,8 @@ ready(() => {
     $loginForm.find('input:visible:first').focus()
   })
 
-  sessionStorage.clear()
+  // do not clear session storage if previewing via the theme editor
+  if (!document.querySelector('.ic-Login--previewing')) {
+    sessionStorage.clear()
+  }
 })
